# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Flyspray",
        "version": "1.0.0",
        "contact": {}
    },
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the user's information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "responses": {
                    "200": {
                        "description": "user is found (User details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Updates the user's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user information",
                "parameters": [
                    {
                        "description": "Update user request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.updateUserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user has been updated successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/refresh-token": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generates a new access token using the provided refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Generate new access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "token has been refreshed successfully (AccessToken \u0026 RefreshToken details are given in a struct in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/signin": {
            "post": {
                "description": "Authenticates a user's credentials and generates access tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Authenticate a user and generate access tokens",
                "parameters": [
                    {
                        "description": "Signin request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.signinBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "logged in successfully (AccessToken details is given in a struct in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Wrong email or password",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "description": "Creates a new user account and sends a verification code to the user's email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user account",
                "parameters": [
                    {
                        "description": "Signup request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.signupBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A verification code has been sent to your email",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "409": {
                        "description": "Email already exists and verified",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/signup/verify": {
            "post": {
                "description": "Verifies a user's account using the provided verification code and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Verify a user's account",
                "parameters": [
                    {
                        "description": "Verification request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.verifyBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "verified",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Email does not exist",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "app.ResponseMsg": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.signinBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "app.signupBody": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "app.updateUserBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "app.verifyBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "verification_code": {
                    "type": "integer"
                }
            }
        }
    }
}